Refactor slider components and update hooks
