+++ components/layout/Slider/MainSlider.module.sass
@@ -1,9 +1,6 @@
 @use '../../../styles/colors' as *
 @use '../../../styles/variables' as *
 
-.container
-  margin: 0
-
 .carouselWrapper
   position: relative
 
@@ -23,8 +20,6 @@
 
 .imgBox
   padding: .5rem 2rem
-  @media screen and (min-width: $small)
-    padding: .5rem 2rem
   @media screen and (min-width: $medium)
     padding: .5rem 10rem
 
@@ -35,9 +30,6 @@
   margin: auto
   height: auto
   box-shadow: var(--img-shadow)
-
-.navButtons
-  display: none
 
 %navBtn
   position: absolute
