+++ components/layout/Slider/ProjectSlider.module.sass
@@ -1,4 +1,3 @@
-@use '../../../styles/colors' as *
 @use '../../../styles/variables' as *
 
 .viewport
